Update index.html
updating current location button. for some reason the pin emoji is now not visible like it was before inside of the button. it is a very cute idea for a current location button. it is a pointed button that spins and pins the current location. when you click on it, it shows the weather for your geolocation. for some reason the button does not display the pin anymore...you can imagine it. or maybe you can see it on your machine, but on mine it does not display...